@@ -1,27 +1,31 @@
 {
   "name": "coqa",
-  "version": "0.1.3",
+  "version": "0.2.0",
   "license": "MIT",
+  "scripts": {
+    "dev": "backpack",
+    "build": "backpack build",
+    "prepublish": "npm run build"
+  },
   "bin": {
-    "coqa": "index.js"
+    "coqa": "./build/main.js"
   },
   "dependencies": {
-    "babel-eslint": "^7.2.3",
+    "babel-eslint": "^8.0.1",
     "babel-plugin-transform-react-jsx": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.26.0",
     "commander": "^2.11.0",
-    "eslint": "^4.5.0",
-    "eslint-plugin-react": "^7.3.0",
+    "eslint": "^4.8.0",
+    "eslint-plugin-react": "^7.4.0",
     "filepaths": "^0.3.0",
     "jsinspect": "^0.12.7",
     "opn": "^5.1.0",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-syntax-highlighter": "^5.7.0",
     "tmp": "^0.0.33"
   },
-  "engines": {
-    "node": ">=7.6.0"
+  "devDependencies": {
+    "backpack-core": "^0.4.3"
   }
 }
